Components refactoring
