--- src/components/film-extra.js
@@ -1,12 +1,0 @@
-const createFilmExtra = () => {
-  return (
-    `<section class="films-list--extra">
-      <h2 class="films-list__title">Top rated</h2>
-      <div class="films-list__container">
-
-      </div>
-    </section>`
-  );
-};
-
-export  {createFilmExtra};
