--- src/components/films.js
@@ -1,14 +1,0 @@
-const createFilms = () => {
-  return (
-    `<section class="films">
-      <section class="films-list">
-        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-        <div class="films-list__container">
-
-        </div>
-      </section>
-    </section>`
-  );
-};
-
-export {createFilms};
